+++ dwds/CHANGELOG.md
@@ -9,6 +9,8 @@
   a given location.
 - Support Chrome `skipLists` to improve stepping performance.
 - Export `AbsoluteImportUriException`.
+- Depend on the latest `package:vm_service` which supports a new
+  `limit` parameter to `getStack`.
 
 ## 7.0.2